server/requirements.txt
 django
 djangorestframework
 requests
 pyyaml
 django-cors-headers
 djangorestframework-simplejwt
 psycopg2
 Pillow
 django-allauth